+++ ChangeLog.md
@@ -9,6 +9,8 @@
   * Control maintenance page through ey.yml options:
   * * maintenance\_on\_restart: true or false (default: false except for glassfish and mongrel)
   * * maintenance\_on\_migrate: true or false (default: true)
+  * Control asset precompilation through ey.yml option:
+  * * precompile\_assets: true or false (default: nil - means assets compilation will be inferred by app/assets and config/application.rb)
 
 ## v1.6.4 (2012-04-26)